# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yield AI API Documentation",
        "description": "Documentation de l'API de prédiction de rendement agricole",
        "contact": {
            "email": "admin@yield.ai"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/predict-yield": {
            "post": {
                "tags": [
                    "Predictions"
                ],
                "summary": "Lancer une prédiction de rendement",
                "operationId": "1e7e7f9c61477ce50a5cc4c96e6e754c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "variete",
                                    "current_week_data"
                                ],
                                "properties": {
                                    "variete": {
                                        "type": "string",
                                        "example": "Variété A"
                                    },
                                    "current_week_data": {
                                        "description": "Données climatiques de la semaine",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Prédiction réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur service IA"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/predictions": {
            "get": {
                "tags": [
                    "Predictions"
                ],
                "summary": "Récupérer l'historique des prédictions",
                "operationId": "a43ff4953f3057eaa399ec5be9eefdf3",
                "responses": {
                    "200": {
                        "description": "Liste des prédictions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "history": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/yield/compare": {
            "post": {
                "tags": [
                    "Analyse de Rendement"
                ],
                "summary": "Comparer prévisions IA et données réelles",
                "description": "Envoie les données réelles au service Python, calcule les métriques (MAE, RMSE) et enregistre le résultat.",
                "operationId": "0feba9f9cc9443af5d0966a34f59da80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prediction_id",
                                    "valeurs_reelles"
                                ],
                                "properties": {
                                    "prediction_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019b6fe7-f207-73c5-b5bc-8fc776729eb5"
                                    },
                                    "valeurs_reelles": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "semaine": {
                                                    "type": "integer",
                                                    "example": 49
                                                },
                                                "valeur_reelle": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 7.9
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Analyse enregistrée ou mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Analyse enregistrée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides ou aucune correspondance trouvée"
                    },
                    "500": {
                        "description": "Erreur interne ou service IA indisponible"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/yield/comparisons": {
            "get": {
                "tags": [
                    "Analyse de Rendement"
                ],
                "summary": "Lister les analyses de l'utilisateur",
                "description": "Récupère toutes les comparaisons de rendement liées aux prédictions de l'utilisateur connecté.",
                "operationId": "495bdaa6b97be51b2f38b7d59182162e",
                "responses": {
                    "200": {
                        "description": "Liste des analyses récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token d'accès (access_token) reçu lors de la connexion pour accéder aux routes protégées.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}